file: config, filter 분리
