+++ src/main/kotlin/com/cymply/auth/config/OAuth2FailureHandler.kt
@@ -1,0 +1,18 @@
+package com.cymply.auth.config
+
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
+import org.springframework.http.HttpStatus
+import org.springframework.security.core.AuthenticationException
+import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler
+import org.springframework.stereotype.Component
+
+@Component
+class OAuth2FailureHandler(
+) : SimpleUrlAuthenticationFailureHandler() {
+    override fun onAuthenticationFailure(
+        request: HttpServletRequest, response: HttpServletResponse, exception: AuthenticationException?
+    ) {
+        response.status = HttpStatus.UNAUTHORIZED.value()
+    }
+}
